refactor #247: refactor utils imports
